[kernel][arch] reset unsafe stack to the top before entering user space
On x86:
Previously, was not resetting the unsafe pointer when entering user
space for the first time, which was apparently leaking a bit less than
200 bytes of the stack which never get reclaimed. Subsequent entry/exits
from user space would never completely unwind the unsafe stack.

On arm64:
Not resetting the base of the shadow call stack on user space entry,
which could waste a few bytes of the shadow call stack. It's not exactly
the same situation as on x86 because the shadow call stack pointer (x18)
is restored to the saved copy in the arch portion of the thread
structure on kernel entry. This copy was only ever set on initial thread
creation and context switch, so it means that it'd only 'leak' the
memory if the thread was context switched at least once prior to exiting
the kernel for the first time.

TODO: consider on both arches just unilaterially reloading the
unsafe/shadow call stack to the base on every entry from user space.
Should be fairly easy to do on arm, but less clean on x86. May or may
not make things more ergonomic however.

Change-Id: I8008c48516261234e106e2bef895de4f913851f9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/523381
Commit-Queue: Travis Geiselbrecht <travisg@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>