[cm][routing] Make check_use_exposed_dir routing tests generic
Adds a method to the RoutingTestModel trait that checks a use
of a capability from an exposed directory and implements the
corresponding tests for both ComponentManager and for the
static routing analyzer.

Also adds a method to ComponentModelForAnalyzer to check a
capability use from an expose declaration when supported for
the capability type.

Bug: 61861
Change-Id: I9a005c68d08aa338b41871296def16407030c39c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/524982
Commit-Queue: Laura Peskin <pesk@google.com>
Reviewed-by: Derek Gonyeo <dgonyeo@google.com>