[netstack] Remove hardError
The purpose of this structure is to leak errors out of loop{Read,Write};
it has hardly anything to do with being "hard". It is also incorrect;
these errors are expected to be consumed after being returned to the
user, which was not being done before this change.

Change-Id: I73dfd6fc1c694387d5b28b9d6f2b6d9936470173
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/525351
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Tamir Duberstein <tamird@google.com>
Reviewed-by: Arthur Sfez <asfez@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Reviewed-by: Mithun Iyer <iyerm@google.com>