[scenic] Allocate protected swapchain buffers early
This CL moves protected buffer allocation to the initialization phase
instead of lazy allocation when needed. This memory comes from a
limit and dedicated pool. We expect to run out of this pool for some
apps. We can avoid allocation failures on the Scenic side by doing it
earlier.

Fixed: 72831, 35785
Change-Id: If058e52dfc8a3a93647e385f454430b473ae2390
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/525045
Commit-Queue: Emircan Uysaler <emircan@google.com>
Reviewed-by: Randall Bosetti <rlb@google.com>