[testing] Add proc macro to help with test fixtures
The #[fixture] proc macro takes one argument, which is the identifier of
a wrapper function.

The wrapper function takes two arguments: the name of the test function,
to be used or ignored however it wants, and a Fn of a signature matching
the input to the proc macro (the test body). The helper is responsible
for doing any necessary setup, and passing inputs to the test body
function. The macro essentially stuffs the test body into a closure,
which is passed to the test wrapper function along with the name of the
test.

For example, this test case:
```
async fn my_test() {
    let (proxy, service_fut) = setup_my_service("my_test");
    let ((), ()) = futures::future::join(service_fut, || async move {
        // test body using `proxy`
    }).await;
}
```

can be rewritten as:
```
async fn my_test(proxy: MyServiceProxy) {
    // test body using `proxy`
}
```

with this helper:
```
async fn setup_my_service<F, Fut>(test_name: &str, test: F)
where
    F: FnOnce(MyServiceProxy) -> impl futures::Future<Output = ()>,
{
    let (proxy, fut) = setup_my_service(test_name);
    let ((), ()) = futures::future::join(fut, test(proxy)).await;
}
```

Bug: 75547
Test: fx test -d fixture-tests
Change-Id: Ib657adc5a70ab37b3dfafc90f204e1ca06e7a388
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/521042
Reviewed-by: James Robinson <jamesr@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Reviewed-by: Tamir Duberstein <tamird@google.com>
Commit-Queue: Peter Johnston <peterjohnston@google.com>