[sdktools] Updates to make SDK tools test friendly
This addresses two issues found when writing tests.

First, on some test environments, the SDK tools are
hermitcally added to the test by using symlinks to the
specific tool. As a result the normal path navigation within the
SDK does not work, and the real path of the executable should be
used to determine the location of other tools.

Second, added a new initializer to fconfig which allows for the
specification of the data directory vs. constructing it from the
user's HOME directory.

Change-Id: If8596ad239d7f32c74cdd04a69d49b8ee32ab38b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/525144
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>
Reviewed-by: Omer Lev-ran <omerlevran@google.com>