[vfs][rust] Update mtime on write calls
Adds a |may_defer| flag to set_attrs which allows filesystems to cache
an attribute change for later flushing.  Once we have writeback caching,
this interface can be used to update the modification time of the file
when we cache the write, without requiring an expensive flush of the
attributes immediately (the filesystem will flush the attributes when it
later actually performs the write).

For now, every write updates mtime.

For fatfs, in which the third-party library manages mtimes itself, the
deferred set_attrs calls are ignored.  This does mean that fatfs is not
quite up to spec (it will not immediately show the new mtime after a
write), and we'll have to go back and fix this later when we implement
writeback caching, but currently it should have no visible effects.

Change-Id: I487267118bc59ff827943b0a09432a71d31fdc30
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/525120
Commit-Queue: James Sullivan <jfsulliv@google.com>
Reviewed-by: Chris Suter <csuter@google.com>