[blobfs] Add host support for reading multi-extent blobs.
This more closely unifies how the host and target read blocks for blobfs
but required some changing of the cache behaviour. The nodemap is now
held in memory as is done on the host so that it can use the same
libraries. This also meant changing the use of a dirty cache block for
flushing the inode. Instead the relevent block is just flushed directly
from the in-memory version, and since this was the only usage of the
dirty cache flag, it has been removed.

Tested: Manually.

Bug: 48571, 32953
Change-Id: I0c10377eb4c4c2fe45156b170286d192835fba12
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/524921
Commit-Queue: Martin Lindsay <mlindsay@google.com>
Reviewed-by: James Sullivan <jfsulliv@google.com>