[debugger] Implement DW_FORM_implicit_const.
This is a new DWARF 5 tag. Currently it seems it is only used by GCC
which is why we had not encountered it earlier.

This was used to encode the size of an enum value, as well as for
DW_AT_decl_file attributes. This attribute is handled by LLVM at a
higher level than we use.

Handle this form in the DwarfDieDecoder by passing the form constant to
the form value.

The above change bypasses the setter for the DWARFUnit that we were
relying on in DWARFFormValue so the unit now needs to be passed around
in more places. This is probably for the better, though, because it
allows us to remove some const_cast calls and lambda captures.

Bug: 76241
Change-Id: I452c33d161993cb4bcc2aa041d1204a10a91e9e3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/526838
Commit-Queue: Brett Wilson <brettw@google.com>
Reviewed-by: Dangyi Liu <dangyi@google.com>