[devices] Ignore peer closed errors in the DriverHost
When the DriverRunner closes the Driver channel, the DriverHost should
not report an error.

I was investigating whether there's a better way to do this, but there
doesn't seem to be a way to inform the server binding that this was
expected. In the other direction, server to client, you can send an
epitaph.

Bug: 33183
Change-Id: Id6b10757d0c783f0cb549b0ab115b1a1ae295138
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/526548
Commit-Queue: Abdulla Kamar <abdulla@google.com>
Fuchsia-Auto-Submit: Abdulla Kamar <abdulla@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>