[blobfs] Report blobs as V_IXUSR
Currently, we report that blobs are regular files (V_TYPE_FILE) and that
they can be read (V_IRUSR). Those are accurate attributes, but it's also
possible that blobs can be executed, which means we should report
V_IXUSR as well.

These attributes are somewhat of a mis-match for our overall system
design because they apply to the blob itself, regardless of how the
client has connected to the blob. In our system, the operations you can
perform on the blob depend on how the client connected.

This situation is much cleaner in fuchsia.io2, which has a different
attribute system that accounts for our Fuchsia actually works.

In the meantime, this CL is helpful because it tells the Android sh
program running in Starnix that it can execute a child process (e.g.,
ls). Without this patch, sh calls stat() on /bin/ls, sees that V_IXUSR
is missing, and refuses to try to run /bin/ls.

Bug: 74060
Change-Id: I0b338382af1b0034fc23b47f9b397c3c682590ec
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/526826
Reviewed-by: Brett Wilson <brettw@google.com>
Commit-Queue: Adam Barth <abarth@google.com>