[multiboot] Don't leave booted (first) kernel item in data ZBI
The Multiboot shim takes a bootable ZBI and peels off the kernel
image from the first item.  Heretofore it then passed the entire
original ZBI as the data ZBI, so the kernel got a copy of its own
image as a data item.  This is confusing to carefully constructed
ZBI tests that expect to find other `KERNEL_*` items that really
are data items.  It's also not adhering to the ZBI boot protocol
as specified, which says that the boot loader delivers a data ZBI
including all the items *after* the initial `KERNEL_*` item, not
*including* it.  So change it to elide that item from the data
ZBI.  To keep it simple and not perturb the overall alignment of
the data ZBI, just edit the item header to replace the type with
ZBI_TYPE_DISCARD and leave it in place.

Bug: 32255
Change-Id: I402e812c33b83a73834e3509561ddb9caf4af65a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/526697
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Roland McGrath <mcgrathr@google.com>
Reviewed-by: Joshua Seaton <joshuaseaton@google.com>