[kernel][ktl] Use a ktl-specific <__config_site> header
Use a local header to override the libc++ installation.  This
removes the need for a config() using defines for ktl.  It also
makes `#include <__config_site>` work again when the toolchain
installation changes so that its version is in a triple-based
subdirectory that isn't present for the modified triple produced
by -m32 for the kernel.phys32 environment.

Bug: 75276
Change-Id: I6308502c5b7d87cba4390907951b0aa8804342d3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/527080
Fuchsia-Auto-Submit: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Petr Hosek <phosek@google.com>