[virtualization] Use more robust test timeouts.
We currently use timeouts on various virtualization guest operations,
such as connecting to guest serial or executing a command over serial.

One difficulty we have with the current set up is that operations with
an individual timeout (for example, running a single command) are used
in loops. This results in either:

1. We need a small command timeout (so the loop doesn't take too
   long), risking test flakes if the command takes longer than
   expected; or

2. We have a long command time with timeout, but now the outer
   loop never terminates because the sum of timeouts is too high.

We've seen several tests failing in CI/CQ with hangs, and it being
unclear what part of a complex operation is going on.

This CL updates the EnclosedGuest (and some of its dependencies) to
support deadlines being passed in. Instead of per-operation timeouts, we
use a single test-wide deadline created at test start time, and pass it
through. Retry loops inside the test framework use their own shorter
deadlines as appropriate, while still respecting the passed-in deadline.

By having more reliable timeouts, this will (hopefully) assist in
diagnosing the reason for hangs.

Multiply: virtualization_core_tests
Bug: 56119
Change-Id: I86b65fffe459e3ddadfd9454893297db72af55d8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/526543
Reviewed-by: Abdulla Kamar <abdulla@google.com>
Commit-Queue: David Greenaway <dgreenaway@google.com>