[virtualization] Don't log everything twice on test failure.
The Logger class supports printing all guest output immediately to
screen with a compile-time flag.

If the flag is enabled and a test is failed, we end up printing the
guest's output twice: once when we first got it, and again after test
failure. This can be confusing when trying to read test output.

Ensure the output is only logged once.

Multiply: virtualization_core_tests
Bug: 56119
Change-Id: I760f6f15669d82cefba2bcae97fd29d4bcacf242
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/526544
Commit-Queue: David Greenaway <dgreenaway@google.com>
Reviewed-by: Abdulla Kamar <abdulla@google.com>