[e2e-tests] Remove the update e2e test
Justification for removal: we have coverage of the same functionality
in other e2e tests, so this has become maintenance burden.

This update test in fact only tests that the 'update' binary functions
and talks to the update installer, which we have coverage of in our
other e2e OTA tests.

Bug: 76034
Change-Id: I247dadd9b59ba981fc17f9417916b974f41f83cf
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/525044
Reviewed-by: Aaron Wood <aaronwood@google.com>
Commit-Queue: John Wittrock <wittrock@google.com>