[fdio] Avoid error-tracking side effects in wait_*
The key insight here is that stream sockets are always closed when they
enter an error state, and that error states are never cleared.

When we observe PEER_CLOSED, use
`fuchsia.posix.socket/BaseSocket.GetError` to retrieve and consume the
error, replacing the previous strategy of encoding the error in the
local state machine. This correctly interacts with `getsockopt(..,
SO_ERROR, ..)` which is also expected to consume the error (and return
no error if the error was already consumed by I/O).  This also correctly
avoids multiple clones of the same socket from all observing the same
error; the first one wins.

The signals previously used to drive the local state machine remain in
use for ABI compatibility until downstream projects are updated to
include the new code.

This makes the situation described in issue 67465 worse because the same
racy side effects now exist in read and write methods. I'm not sure how
we can possibly avoid this without changing zircon sockets to start out
closed so that they can be opened from the netstack upon connection
establishment.

Bug: 67465
Bug: 75717
Change-Id: Ie7c9c9ff2493b4ec3b3b4ccbc2761b54b6eb6bbb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/526061
Commit-Queue: Tamir Duberstein <tamird@google.com>
Reviewed-by: Mithun Iyer <iyerm@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>