[bt][a2dp] Add v2 manifest
Add a CFv2 manifest for the A2DP component as well as an A2DP manifest
shard to be used by integration tests. This simplifies the smoke-test to
verify capability routing WAI and not component behavior. Defer the
declaration of the child component to the manifest so that tests can
hermetically include their own version of the child component.

Bug: 73935
Test: fx test bt-a2dp-smoke-test

Change-Id: I4a612ac38077c8cca5236ee3cf150cb853a16028
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/521863
Commit-Queue: Ani Ramakrishnan <aniramakri@google.com>
Fuchsia-Auto-Submit: Ani Ramakrishnan <aniramakri@google.com>
Reviewed-by: Lucas Jenkins <lucasjenkins@google.com>