[sysmem] Fix leak caused by GetAuxBuffers
Sysmem.HeapAmlogicSecure uses
fuchsia_sysmem_BufferCollectionGetAuxBuffers, which does
BufferCollection::CloneResultForSendingV2. The new collection info is
allocated in the TableSet, so even when it goes out of scope the VMOs
it references may still be held onto. This causes a reference cycle
to the LogicalBufferCollection, so it was leaked forever.

If we recreate the TableSet allocator in FailDownFrom we can ensure that
the VMO handles will be destroyed and the reference cycle will be
avoided.

Change-Id: Ib99b45012e85d8de5517f2335189e8ff239af0c6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/526300
Commit-Queue: John Bauman <jbauman@google.com>
Reviewed-by: Dustin Green <dustingreen@google.com>