[component_id_index] Fix monikers with nonzero instance IDs being rejected
The component ID index `moniker` field does not accept instance
identifiers. When the moniker string is parsed, it is converted to
an AbsoluteMoniker comprised of ChildMonikers that all have the
instance ID set to 0.

The check fails if the moniker of the incoming component has a
non-zero instance ID. e.g. with the index entry:

    {
        instance_id: "...",
        moniker: "/core/session-manager/session:session"
    }

component_manager returns the error:

    [component_manager] ERROR: bind() failed: Routing error:
    Component /core:0/session-manager:0/session:session:1 is trying
    to use a storage capability which is restricted to the
    component ID index.

Change-Id: I610b26b8a28076500f3c55eb1aab42abf0ee7be7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/525728
Reviewed-by: Gary Bressler <geb@google.com>
Reviewed-by: Vardhan Mudunuru <vardhan@google.com>
Commit-Queue: Yegor Pomortsev <ypomortsev@google.com>