[fidlc] Use separate root typespaces for old/new syntax
This CL supports the use of client_end/server_end in the
new syntax (and disallows request in the old syntax) by
having two different root typespaces that are used
depending on the syntax.

In the process, it adds a new type template and type
pair (TransportSideTypeTemplate & TransportSideType) to
represent client and server ends in the new syntax.

Bug: 70186
Test: fidldev test fidlc
Change-Id: Ifcd9a7128f4bbd08a5de45a4a1342bc11cab04f0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/487181
Commit-Queue: Felix Zhu <fcz@google.com>
Reviewed-by: Alex Zaslavsky <azaslavsky@google.com>