[image assembly][build] Add support for json vbmeta descriptor files
Add support in the vbmeta signing template to take a list of json files
that describe the hash descriptors to add to the vbmeta image, instead
of vbmeta images that needs to have the descriptors extracted from.

This is a part of a soft transition.  The next step here will be to
remove support for the vbmeta hash descriptor extraction.

Change-Id: Ib21bd75f6d03852e17d4d0c2480646d58ef90028
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/526580
Commit-Queue: Aaron Wood <aaronwood@google.com>
Reviewed-by: Allison Pearce <ampearce@google.com>
Reviewed-by: Aidan Wolter <awolter@google.com>
Reviewed-by: David Pursell <dpursell@google.com>
Reviewed-by: Ali Zhang <alizhang@google.com>