[starnix][loader] Track memory from loading ELFs
This CL causes the MemoryManager to record the maps made when we load an
ELF into the task. Remembering these maps is important for when we fork
a child process so that the child process can have its code mapped into
its address space.

Bug: 76319
Change-Id: If8302c636ddd272cf540f7e0c2d45a46c0d338fa
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/527360
Commit-Queue: Adam Barth <abarth@google.com>
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>