[pkg-resolver] Persist TUF metadata when configured
This change enables support for the pkg-resolver to persist TUF metadata
across boots / restarts when it is configured to do so. An empty-string
or absent configuration prevents this behavior from having effect. This
configuration is not enabled in any product builds by default.

Test environment updated to allow configuring persisted
ServedRepositories.

Test: new tests for single successful configuration setting, three
failure cases, and on-device test with the target included in fx set.

Fixed: 73106

Change-Id: I65b2918a9bcba3f6de4f3ce32de3d0068a8ea9f7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/524342
Commit-Queue: Devon H. O'Dell <dhobsd@google.com>
Reviewed-by: Allison Pearce <ampearce@google.com>
Reviewed-by: Zach Kirschenbaum <zkbaum@google.com>