[cmpctmalloc] Decouple max grow and alloc limits
This change rearranges some definitions with end result that the amount
the heap is able to grow by can more than the largest allocation that
it would actually satisfy.

In doing this various comments and intermediate definitions are added to
provide clarity due to internal metadata headers being included or not
included at various points in different calculations. This should
ensure that when the values for the maximum allocation and grow sizes
are changed in the future that incorrect/incompatible values are less
likely to be chosen.

Bug: 30973
Change-Id: I438a6336801e3fe525638e39fce552e47b868881
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/519890
Fuchsia-Auto-Submit: Adrian Danis <adanis@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>