[carnelian] Add layout gallery
Add an example of layouts that also acts as layout unit tests. This
setup allows a test author to eyeball the layout produced by the test
and, if acceptable, capture the resulting facet bounds for use in a
test.

As part of this change, actually implement the remaining 8 stack
alignments that were omitted in the original CL.

In order to serialize euclid rectangles, turn on the serde feature of
euclid.

Also include a file fx format-code touched.

Testing: unit tests plus manual.

Change-Id: I59462a6eb45bc15bc2ffb1ae55d21b734f7df6ab
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/523285
Reviewed-by: David Reveman <reveman@google.com>
Reviewed-by: Adam Perry <adamperry@google.com>
Commit-Queue: Rob Tsuk <robtsuk@google.com>