[build] cargo_toml_gen has non-hermetic deps
//build/rust:cargo_toml_gen is used to generate Cargo.toml files based
on the GN build graph. It uses GN which reads files throughout the
source tree, so it is a non-hermetic action. Since this is only used
locally by developers it should be fine for these to be non-hermetic
long term without affecting incremental builds in CQ.

This allows for using `build_should_trace_actions=true` locally if using
:cargo_toml_gen.

Change-Id: If0152b5fb55324ced534469848a1b7ffe070b7fb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/528291
Reviewed-by: Shai Barack <shayba@google.com>
Commit-Queue: Bryan Henry <bryanhenry@google.com>