[debugger] Use a struct for process/thread koid.
Longer-term I want to make it easy to take a list of threads from the
debug agent that stopped and resume them. This was challenging with the
old setups, which for example had a process koid and then a vector of
thread koids in the resume request.

This adds a process/thread koid pair struct and uses it in the places in
the IPC protocol where appropriate. Also use it in some associated
non-IPC code.

Fix a problem I noticed in DebugAgent::OnResume where individual thread
IDs were ignored.

Change-Id: Ifbfc20706854614444bcf2ffeed7f837dd9ea1bb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/525882
Reviewed-by: Vincent Belliard <vbelliard@google.com>
Reviewed-by: Michael Jones <michaelrj@google.com>
Reviewed-by: Dangyi Liu <dangyi@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>