[repro] Copy fx repro script into fuchsia.git
It was previously only available in internal checkouts because it used
to rely on `fx sync-to`, which is internal, but `fx sync-to` is now
optional when using `fx repro`, and nothing else in `fx repro` depends
on internal code.

The only changes I made to the script were to add the check to see
whether the `sync-to` command is available, as well as to add a `set
-euf -o pipefail` line and fixing a few broken pipes and unbound
variables.

See http://go/paste/6517091814342656 for the diff.

Bug: 74724
Change-Id: I41d79c691e50566dfbbb1a5dd3d30d1e921e43cd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/528540
Fuchsia-Auto-Submit: Oliver Newman <olivernewman@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Gary Boone <gboone@google.com>