[assembly] Separate product and board configs
The product and board fields are being separated into different configs.
In the future we may want to re-evaluate this, but for now it is
consistent with the current assembly process.

I filed fxb/76370 to track re-evaluating this.

Bug: 76379
Change-Id: Iefba8e5fb42773b9899b95760f90ea0c8fc6a88e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/527722
Reviewed-by: Aaron Wood <aaronwood@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Aidan Wolter <awolter@google.com>