[ffx] fix timeout and port bugs in get-ssh-address
The timeout now allows for the target to be discovered after the command
is started. The output of the command always includes a port number,
defaulting to port 22 if none is set on the target.  Minor updates to
the error output string.

Fixed: 75938
Change-Id: If7f04bfe5d77395de920adef593ffe26230f79eb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/526944
Fuchsia-Auto-Submit: James Tucker <raggi@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Andy Weiss <dragonbear@google.com>