[ime_service] Simplify keyboard focus events
Before:
The keyboard service tracks, for each subscriber, which keys had been
depressed when the subscriber lost focus. When it next regains focus,
the subscriber receives CANCEL events for keys it was tracking that
have since been released, and SYNC events only for newly pressed keys
that it wasn't tracking.
The subscriber is led to believe that the keys had remained pressed the
entire time that it was unfocused.

After:
When a subscriber loses focus, it immediately receives CANCEL events
for any currently depressed keys. When a subscriber gains focus, it
receives SYNC events for all currently depressed keys. There is no per-
subscriber tracking of keys. From the the subscriber's point of view,
no keys are ever in a depressed state while the subscriber is
unfocused.

TESTING:
1. Include `--with src/ui/bin/ime:tests` in fx set args.
2. Run `fx test keyboard3_integration_test`.

Bug: 76086
Change-Id: I6b90cc3459c0b748bd852c76229f508590f58c46
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/526507
Reviewed-by: Filip Filmar <fmil@google.com>
Commit-Queue: Konstantin Pozin <kpozin@google.com>