[fidl] Update DriverIndex.MatchDriver to support composites
This allows the driver index to provide the driver runner with
information about when a node matches a composite driver.

We use the `node_id` to specify which node was matched, and `num_nodes`
to specify the total number of nodes to be matched. This way, the driver
index does not need to hold any state, and the driver runner can make
the decision about when to create a composite node for a composite
driver.

Bug: 73162
Bug: 33183
Change-Id: I0a8cb2d4451e6984380db072afaa25bc2a94ae6e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/528700
Reviewed-by: David Gilhooley <dgilhooley@google.com>
API-Review: Jocelyn Dang <jocelyndang@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Abdulla Kamar <abdulla@google.com>