[build][fidl] Add fidlgentest_go_test template
This CL adds a template for defining a go_test target that uses the
fidlgentest library, taking care of recording the dependency on fidlc
and passing it as an argument to the test binary.

This reduces duplication and makes in harder to use fidlgentest wrong.
In particular, it used to be easy to forget the host_test_data step for
fidlc, leading to flaky failures that only show up in CQ. This had to be
fixed twice recently:

* If370d7b7e49ddd1f12531bd1575cde40c61cbfdc
* I046db8071a8de004feb61585e6723ff03e111b58

Test: fx test fidlgen summarize apidiff
Change-Id: Ic62002cae6850c053b69ab42571545679513acc1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/525799
Commit-Queue: Mitchell Kember <mkember@google.com>
Reviewed-by: Pascal-Louis Perez <pascallouis@google.com>
Reviewed-by: Ian McKellar <ianloic@google.com>