[starnix_test_runner] Keep component controller alive
This change keeps the component controller for the test component alive
until there are no more test suite requests for that component.

It also only serves one SuiteRequestStream, since it needs to close the
component controller on completion.

This is an attempt to fix a flake where a protocol is failed to be
routed, sinced the component "instance is not running."

Bug: 76256
Change-Id: Ia1ff1390f4ca857f9a7b05af8532b54924ba896d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/528920
Commit-Queue: Kevin Lindkvist <lindkvist@google.com>
Reviewed-by: Adam Barth <abarth@google.com>