[logs] Add support for JSON namespaces
In order to support future extensions to structured logs,
we need to add namespaces to our logs. This is a breaking
format change that creates a new "message" namespace used
for encoding regular log messages. In the future,
"printf" will be used for structured printf encoding,
and "graph" will be used for graph encoding.

Change-Id: I2788c17aa125e9f34da851a9ce6372f1d338c4e1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/522140
Reviewed-by: Christopher Johnson <crjohns@google.com>
Reviewed-by: Miguel Flores <miguelfrde@google.com>
Reviewed-by: Shai Barack <shayba@google.com>
Commit-Queue: Brian Bosak <bbosak@google.com>