[root-presenter] propagate VKManager notifications to VKController
When a VirtualKeyboardManager notifies the VirtualKeyboardCoordinator
of a change in the visibility of the keyboard, pass that information
on to the VirtualKeyboardController.

Also: add a FIDL test which verifies that a notification triggered by
a user interaction is received by a call to
fuchsia.input.virtualkeyboard.Controller.WatchVisibility().

Along the way, separate VirtualKeyboardController into an interface
and an implementation, to ease unit testing of VirtualKeyboardCoordinator.

Bug: 73279
Bug: 73281
Test: fx test -o root_presenter_unittests  -- --gtest_filter='VirtualKeyboard*'
Change-Id: Iec5ade01a2ad6f88fbefdf7e0b35a716d45ec3c1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/525035
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Mukesh Agrawal <quiche@google.com>
Reviewed-by: Alice Neels <neelsa@google.com>