[image assembly] Start transition for pkgfs allowlist
This starts the transition of the pkgfs packages allowlist from a bare
file that's included via gn args in the system-image package to a
resource() template invocation, and a fuchsia_system_package()
template invocation, which is then added to the
product_system_image_deps group.

1 of 5

Setup the new logic, so that products can set the allowlist arg to empty
if they include a fuchsia_system_package() entry.

Bug: 76460
Change-Id: I2e7b705bde54fb7ac089d6f8d145f9d43b244af0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/528423
Commit-Queue: Aaron Wood <aaronwood@google.com>
Reviewed-by: Dan Johnson <computerdruid@google.com>
Reviewed-by: Allison Pearce <ampearce@google.com>