[root-presenter] propagate VKController requests to VKManager
When a VirtualKeyboardController requests a change in the type or
visibility of the keyboard, pass that information on to the
VirtualKeyboardManager.

Add FIDL tests which exercise the ability to receive this information
via fuchsia.input.virtualkeyboard.Manager.WatchTypeAndVisibility().

Bug: 73279
Bug: 73281
Test: fx test -o root_presenter_unittests  -- --gtest_filter='VirtualKeyboard*'
Change-Id: I91bd0a1a39574a5b5d844be17128028f69c5d1ad
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/525038
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Mukesh Agrawal <quiche@google.com>
Reviewed-by: Alice Neels <neelsa@google.com>