[phys][page-table] Add address space builder for ARM64
Add a MapPage implementation for ARM64 and use this to create an ARM64
version of an AddressSpaceBuilder.

The builder will be used in upcoming CLs to create a 1:1 phys/virt page
table mapping, and longer term to map in the kernel at its final vaddr.

Currently the builder has a few limitations: it can only be used for
the TTBR0 address space (i.e., the low half of virtual memory), and
only supports a hard-coded set of attributes for mappings (in
particular, supervisor-only writability and the 0'th memory attribute).
These will be addressed in upcoming CLs.

Bug: 67632
Change-Id: I26eb0d7b19f7fd61345f24e12c30bec49bea91ee
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/513701
Reviewed-by: Adrian Danis <adanis@google.com>
Reviewed-by: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Fuchsia-Auto-Submit: David Greenaway <dgreenaway@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>