[fidlc] Disallow multiple constraints without braces
Previously this was not allowed, but we would throw an error
telling the user to surround multiple constraints in brackets.
Handling that error specifically is actually not possible since
it leads to a grammatical ambiguity.

This also removes the ErrUnnecessaryConstraintBrackets, which is
allowed according to RFC-0050

Fixed: 76189
Test: fx ninja -C out/default host_x64/fidl-compiler &&
./out/default/host_x64/fidl-compiler

Change-Id: I9361844dae647879618a1510f786bd2ae2f7310a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/529041
Commit-Queue: Felix Zhu <fcz@google.com>
Reviewed-by: Alex Zaslavsky <azaslavsky@google.com>