[input] Adds the text settings input pipeline stage.
This stage is used to decorate the input pipeline with the
currently active keymap.  The keymap configuration is very
minimal at the moment and only allows selecting between two
keymaps.  We expect this to evolve into support for more
keymaps in the medium term.

Also added is a CLI program that allows setting the keymap
from the command line, for the most rudimentary manual
tests.

Fixed: 75722
Change-Id: Id74658a665cbdf7d7ecd077fdecc1d055d747b54
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/529180
Reviewed-by: Mukesh Agrawal <quiche@google.com>
Reviewed-by: Vickie Cheng <vickiecheng@google.com>
Commit-Queue: Filip Filmar <fmil@google.com>