[ffx] get-ssh-address promptly sees targets
Manually added targets were not generating NewTargetEvents due to an
otherwised unused condition that does not fire events if the target is
unnamed.
get-ssh-address would take a minimum of 8s as a result of the poll
missing the event, but the timeout it used from get_target was 8s.

Change-Id: I391f8262f778f7ab9eb822ed0bc793a6b207c370
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/530047
Reviewed-by: Andrew Davies <awdavies@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: James Tucker <raggi@google.com>