[bootstrap] Add DriverIndex routing
Add the DriverIndex routing to bootstrap. DriverIndex will not
exist in most builds because it is currently only needed for
drivers as components, but the routing means that it will be
used correctly for builds that include it.

Change-Id: Ic40c42a9a815625753b01410077789d9491765ca
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/530162
Fuchsia-Auto-Submit: David Gilhooley <dgilhooley@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Abdulla Kamar <abdulla@google.com>