[tools][fuzzing] Use tempfiles for Handle implementation in undercoat
In the previous implementation, the length of the serialized
representation of a Handle was directly proportional to the amount of
data stored in the Handle, resulting in some unwieldy tokens. This
changes the implementation to a more compact, tempfile-based one that
also gives us the benefits of 1) serializing type information so that
multiple types of connectors/launchers/etc can be used, and 2) being
able to store mutable data in a fixed handle if needed in the future.

Also updates a few paths to host tools and fixes a typo in a LSC
associated with the tools/qemu library.

Bug: 47320
Bug: 47479

Change-Id: I26d8aa31cd7406b45f809663645753dad45d28bc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/436748
Commit-Queue: Cameron Finucane <eep@google.com>
Reviewed-by: Aaron Green <aarongreen@google.com>