[jira] Fix additional malformed TODOs.
This is the last of it for Fuchsia, these are TODOs which seemingly use
a Jira project prefix with a valid monorail bug id. I manually verified
that the monorail ids are in fact the correct bug for the context.

Bug: 60389
Change-Id: I06e1b671dfc202514374da29e20cd40026b721ae
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/530147
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Nathan Mulcahey <nmulcahey@google.com>
Reviewed-by: Shai Barack <shayba@google.com>
API-Review: Shai Barack <shayba@google.com>