[archivist] Unified & Feedback reader is a v2 test
As a result a lot of duplicated code is now shared for both tests.

To be able to set /config/data in the test, a new configuration option
is introduced to specify the location of pipeline config. By default
this would be /config/data but for the test purposes it's set to
/pkg/config/data/pipelines.

Bug: 76502
Tested: archivist-integration-tests-v2

Change-Id: Ic1556470bdf2a61d7464301ff9100f59a7a6114f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/530462
Fuchsia-Auto-Submit: Miguel Flores <miguelfrde@google.com>
Reviewed-by: Gary Bressler <geb@google.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>
Commit-Queue: Miguel Flores <miguelfrde@google.com>