[pkg-cache] Perform GC directly in pkg-cache
This moves garbage collection out of pkgfs and into pkg-cache. The
approach currently implemented is intended to be entirely compatible
with pkgfs. The only material difference is that, due to the NeededBlobs
protocol releasing blobs from the dynamic index after a failure,
incomplete blobs may be protected for longer in pkgfs than in pkg-cache.

Test: new tests added verifying parity between pkgfs and SpaceManager.
Fixed: 66886
Fixed: 66888

Change-Id: Ib93d04e5d9e056c600bd3cc79e539e8d943b2bcb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/514646
Commit-Queue: Devon H. O'Dell <dhobsd@google.com>
Reviewed-by: Dan Johnson <computerdruid@google.com>